FT_TRANSCENDANCE/backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.4
django-cors-headers==4.3.1
django-crontab==0.7.1
django-extensions==3.2.3
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
djangorestframework==3.15.1
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.1
fabric==3.2.2
fake==0.8
Faker==25.5.0
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
invoke==2.2.0
msgpack==1.0.8
multidict==6.0.5
paramiko==3.4.0
phonenumbers==8.13.36
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
qrcode==7.4.2
redis==5.0.4
requests==2.31.0
sendgrid==6.11.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
twilio==9.0.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zope.interface==6.4.post2
